pyproject.toml
[project]
name = "trav-planner-assistant"
version = "0.1.0"
description = "Multi-agent system for constraint-aware travel planning and disruption recovery"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0",
    "anthropic>=0.40",
    "langsmith>=0.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "langchain-core>=0.3",
    "langgraph-checkpoint-sqlite>=3.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.0",
]

[project.scripts]
trav-planner = "cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: requires ANTHROPIC_API_KEY (skipped in CI if not set)",
]

[tool.hatch.build.targets.wheel]
packages = ["models", "graph", "agents", "rag"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
